Level 2-9 : ArithmeticCalculator 클래스의 연산 메서드에 책임(역할)이 많아 보입니다. 사칙연산 각각의 기능을 담당하는 AddOperator, SubtractOperator, MultiplyOperator, DivideOperator 클래스를 만들어 연산 메서드의 책임을 분리. (SRP)

+++ src/main/java/calculator/Calculator.java
@@ -36,7 +36,7 @@ public abstract class Calculator {// 연산 결과를 저장하고 있는 컬렉
     }
 
     // 추상 메서드
-    protected abstract double calculate(int firstNumber, int secondNumber, char operator) throws InputErrorException;
+    protected abstract int calculate(int firstNumber, int secondNumber, char operator) throws InputErrorException;
     protected abstract double calculate(int radius);
 
 }
